Create Daniel's main
The top part that is commented is to get the Accelerometer values. The values are somewhat cleaned out, but could maybe use some work.

The bottom code is using the button to start the locking and unlocking sequence. I used button presses for locking and unlocking, but that will be changed with the moving board thing